+++ src/painlessmesh/router.hpp
@@ -195,7 +195,8 @@ size_t broadcast(protocol::Variant&& variant, layout::Layout<T> layout,
 
 template <class T>
 void routePackage(layout::Layout<T> layout, std::shared_ptr<T> connection,
-                  const TSTRING& pkg, callback::MeshPackageCallbackList<T> cbl,
+                  const TSTRING& pkg,
+                  callback::MeshPackageCallbackList<T>& cbl,
                   uint32_t receivedAt) {
   using namespace logger;
   Log(COMMUNICATION, "routePackage(): Recvd from %u: %s\n", connection->nodeId,
@@ -335,8 +336,8 @@ void handleNodeSync(T& mesh, protocol::NodeTree newTree,
 }
 
 template <class T, typename U>
-callback::MeshPackageCallbackList<U> addPackageCallback(
-    callback::MeshPackageCallbackList<U>&& callbackList, T& mesh) {
+void addPackageCallback(callback::MeshPackageCallbackList<U>& callbackList,
+                        T& mesh) {
   // REQUEST type,
   callbackList.onPackage(
       protocol::NODE_SYNC_REQUEST,
@@ -359,8 +360,6 @@ callback::MeshPackageCallbackList<U> addPackageCallback(
         connection->timeOutTask.disable();
         return false;
       });
-
-  return callbackList;
 }
 
 }  // namespace router